chore: custom hook useFilterCountries replaces getCountriesByFilter function
