+++ src/pages/HomePage/HomePage.tsx
@@ -1,29 +1,17 @@
-import { useEffect, useState } from 'react';
+import { useState } from 'react';
 import { CountryCard } from '../../components/CountryCard/CountryCard';
 import { SearchInput } from '../../components/SearchInput/SearchInput';
 import { SearchSelect } from '../../components/SearchSelect/SearchSelect';
 import { CountriesData } from '../../types/CountriesTypes';
-import { getCountriesByFilter } from '../../helpers/getCountriesByFilter';
+import { useFilterCountries } from '../../hooks/useFilterCountries';
 import './HomePageStyles.css';
 
 export const HomePage = () => {
 
   const [countryRegion, setCountryRegion] = useState('');
   const [countryName, setCountryName] = useState('');
-  const [countries, setCountries] = useState<CountriesData[]>([]);
 
-  const getCountriesData = async (region:string, name:string) => {
-    try {
-      const countriesInfo:CountriesData[] = await getCountriesByFilter(region,name);
-      setCountries(countriesInfo);
-    } catch (error) {
-      console.error(error)
-    }
-  }
-
-  useEffect(() => {
-    getCountriesData(countryRegion,countryName);
-  },[countryRegion,countryName]);
+  const countries:CountriesData[] = useFilterCountries(countryRegion,countryName);
   
   return (
     <>
